Refactored to match structures.
